+++ sdk/core/azure_core_opentelemetry/tests/telemetry_service_macros.rs
@@ -685,10 +685,7 @@ mod tests {
         let package_version = env!("CARGO_PKG_VERSION").to_string();
         azure_core_test::tracing::assert_instrumentation_information(
             |tracer_provider| Ok(create_service_client(&ctx, tracer_provider)),
-            |client| {
-                let client = client;
-                Box::pin(async move { client.get("get", None).await })
-            },
+            async move |client| client.get("get", None).await,
             ExpectedInstrumentation {
                 package_name,
                 package_version,
@@ -706,14 +703,11 @@ mod tests {
 
     #[recorded::test()]
     async fn test_function_tracing_tests(ctx: TestContext) -> Result<()> {
-        let package_name = env!("CARGO_PKG_NAME").to_string();
+        let package_name = ctx.recording().var("CARGO_PKG_NAME", None).to_string();
         let package_version = env!("CARGO_PKG_VERSION").to_string();
         azure_core_test::tracing::assert_instrumentation_information(
             |tracer_provider| Ok(create_service_client(&ctx, tracer_provider)),
-            |client| {
-                let client = client;
-                Box::pin(async move { client.get_with_function_tracing("get", None).await })
-            },
+            async move |client| client.get_with_function_tracing("get", None).await,
             ExpectedInstrumentation {
                 package_name,
                 package_version,
@@ -737,14 +731,11 @@ mod tests {
     async fn test_function_tracing_tests_error(ctx: TestContext) -> Result<()> {
         use azure_core_test::tracing::ExpectedRestApiSpan;
 
-        let package_name = env!("CARGO_PKG_NAME").to_string();
+        let package_name = ctx.recording().var("CARGO_PKG_NAME", None).to_string();
         let package_version = env!("CARGO_PKG_VERSION").to_string();
         azure_core_test::tracing::assert_instrumentation_information(
             |tracer_provider| Ok(create_service_client(&ctx, tracer_provider)),
-            |client| {
-                let client = client;
-                Box::pin(async move { client.get_with_function_tracing("index.htm", None).await })
-            },
+            async move |client| client.get_with_function_tracing("index.htm", None).await,
             ExpectedInstrumentation {
                 package_name,
                 package_version,