+++ sdk/keyvault/azure_security_keyvault_secrets/tests/secret_client.rs
@@ -268,6 +268,8 @@ async fn round_trip_secret_verify_telemetry(ctx: TestContext) -> Result<()> {
         },
         ExpectedInstrumentation {
             package_name: recording.var("CARGO_PKG_NAME", None),
+
+            // Don't use `recording.var` here in case the recording was made with a different package version.
             package_version: env!("CARGO_PKG_VERSION").into(),
             package_namespace: Some("KeyVault"),
             api_calls: vec![
@@ -289,4 +291,283 @@ async fn round_trip_secret_verify_telemetry(ctx: TestContext) -> Result<()> {
     .await?;
 
     Ok(())
+}
+
+#[recorded::test]
+async fn list_secrets_verify_telemetry(ctx: TestContext) -> Result<()> {
+    use azure_core_test::tracing::ExpectedRestApiSpan;
+
+    const SECRET_COUNT: usize = 50;
+
+    let recording = ctx.recording();
+
+    {
+        let secret_client = {
+            let mut options = SecretClientOptions::default();
+            recording.instrument(&mut options.client_options);
+            SecretClient::new(
+                recording.var("AZURE_KEYVAULT_URL", None).as_str(),
+                recording.credential(),
+                Some(options),
+            )
+        }?;
+        for i in 0..SECRET_COUNT {
+            let secret = secret_client
+                .set_secret(
+                    &format!("secret-list-telemetry-{}", i),
+                    SetSecretParameters {
+                        value: Some(format!("secret-list-telemetry-value-{}", i)),
+                        ..Default::default()
+                    }
+                    .try_into()?,
+                    None,
+                )
+                .await?
+                .into_model()?;
+            assert_eq!(
+                secret.value,
+                Some(format!("secret-list-telemetry-value-{}", i))
+            );
+        }
+    }
+    // Verify that the distributed tracing traces generated from the API call below match the expected traces.
+    let validate_result = azure_core_test::tracing::assert_instrumentation_information(
+        |tracer_provider| {
+            let mut options = SecretClientOptions::default();
+            recording.instrument(&mut options.client_options);
+            options.client_options.instrumentation = InstrumentationOptions {
+                tracer_provider: Some(tracer_provider),
+            };
+            SecretClient::new(
+                recording.var("AZURE_KEYVAULT_URL", None).as_str(),
+                recording.credential(),
+                Some(options),
+            )
+        },
+        async move |client: SecretClient| {
+            let mut secrets = client.list_secret_properties(None)?;
+            while let Some(secret) = secrets.try_next().await? {
+                let _ = secret.resource_id()?;
+            }
+
+            Ok(())
+        },
+        ExpectedInstrumentation {
+            package_name: recording.var("CARGO_PKG_NAME", None),
+            // Don't use `recording.var` here in case the recording was made with a different package version.
+            package_version: env!("CARGO_PKG_VERSION").into(),
+            package_namespace: Some("KeyVault"),
+            api_calls: vec![ExpectedApiInformation {
+                api_name: Some("KeyVault.getSecrets"),
+                api_children: vec![ExpectedRestApiSpan {
+                    api_verb: azure_core::http::Method::Get,
+                    is_wildcard: true,
+                    ..Default::default()
+                }],
+                ..Default::default()
+            }],
+        },
+    )
+    .await;
+
+    validate_result
+}
+
+#[recorded::test]
+async fn list_secrets_by_pages_verify_telemetry(ctx: TestContext) -> Result<()> {
+    use azure_core_test::tracing::ExpectedRestApiSpan;
+
+    const SECRET_COUNT: usize = 50;
+
+    let recording = ctx.recording();
+
+    {
+        let secret_client = {
+            let mut options = SecretClientOptions::default();
+            recording.instrument(&mut options.client_options);
+            SecretClient::new(
+                recording.var("AZURE_KEYVAULT_URL", None).as_str(),
+                recording.credential(),
+                Some(options),
+            )
+        }?;
+        for i in 0..SECRET_COUNT {
+            let secret = secret_client
+                .set_secret(
+                    &format!("secret-list-telemetry-by-page{}", i),
+                    SetSecretParameters {
+                        value: Some(format!("secret-list-telemetry-by-page-value-{}", i)),
+                        ..Default::default()
+                    }
+                    .try_into()?,
+                    None,
+                )
+                .await?
+                .into_model()?;
+            assert_eq!(
+                secret.value,
+                Some(format!("secret-list-telemetry-by-page-value-{}", i))
+            );
+        }
+    }
+    // Verify that the distributed tracing traces generated from the API call below match the expected traces.
+    let validate_result = azure_core_test::tracing::assert_instrumentation_information(
+        |tracer_provider| {
+            let mut options = SecretClientOptions::default();
+            recording.instrument(&mut options.client_options);
+            options.client_options.instrumentation = InstrumentationOptions {
+                tracer_provider: Some(tracer_provider),
+            };
+            SecretClient::new(
+                recording.var("AZURE_KEYVAULT_URL", None).as_str(),
+                recording.credential(),
+                Some(options),
+            )
+        },
+        async move |client: SecretClient| {
+            let mut secrets = client.list_secret_properties(None)?.into_pages();
+            while let Some(page) = secrets.try_next().await? {
+                let items = page.into_model()?;
+                for item in items.value {
+                    let _ = item.resource_id()?;
+                }
+            }
+
+            Ok(())
+        },
+        ExpectedInstrumentation {
+            package_name: recording.var("CARGO_PKG_NAME", None),
+            // Don't use `recording.var` here in case the recording was made with a different package version.
+            package_version: env!("CARGO_PKG_VERSION").into(),
+            package_namespace: Some("KeyVault"),
+            api_calls: vec![ExpectedApiInformation {
+                api_name: Some("KeyVault.getSecrets"),
+                api_children: vec![ExpectedRestApiSpan {
+                    api_verb: azure_core::http::Method::Get,
+                    is_wildcard: true,
+                    ..Default::default()
+                }],
+                ..Default::default()
+            }],
+        },
+    )
+    .await;
+
+    validate_result
+}
+
+#[recorded::test]
+async fn list_secrets_verify_telemetry_rehydrated(ctx: TestContext) -> Result<()> {
+    use azure_core_test::tracing::ExpectedRestApiSpan;
+
+    const SECRET_COUNT: usize = 50;
+
+    let recording = ctx.recording();
+
+    {
+        let secret_client = {
+            let mut options = SecretClientOptions::default();
+            recording.instrument(&mut options.client_options);
+            SecretClient::new(
+                recording.var("AZURE_KEYVAULT_URL", None).as_str(),
+                recording.credential(),
+                Some(options),
+            )
+        }?;
+        for i in 0..SECRET_COUNT {
+            let secret = secret_client
+                .set_secret(
+                    &format!("secret-rehydrate-telemetry-{}", i),
+                    SetSecretParameters {
+                        value: Some(format!("secret-rehydrate-telemetry-value-{}", i)),
+                        ..Default::default()
+                    }
+                    .try_into()?,
+                    None,
+                )
+                .await?
+                .into_model()?;
+            assert_eq!(
+                secret.value,
+                Some(format!("secret-rehydrate-telemetry-value-{}", i))
+            );
+        }
+    }
+    // Verify that the distributed tracing traces generated from the API call below match the expected traces.
+    let validate_result = azure_core_test::tracing::assert_instrumentation_information(
+        |tracer_provider| {
+            let mut options = SecretClientOptions::default();
+            recording.instrument(&mut options.client_options);
+            options.client_options.instrumentation = InstrumentationOptions {
+                tracer_provider: Some(tracer_provider),
+            };
+            SecretClient::new(
+                recording.var("AZURE_KEYVAULT_URL", None).as_str(),
+                recording.credential(),
+                Some(options),
+            )
+        },
+        async move |client: SecretClient| {
+            let rehydration_token = {
+                let mut first_pager = client.list_secret_properties(None)?.into_pages();
+
+                // Prime the iteration.
+                let first_page = first_pager
+                    .try_next()
+                    .await?
+                    .expect("expected at least one page");
+                {
+                    let secrets = first_page.into_model()?;
+                    for secret in secrets.value {
+                        let _ = secret.resource_id()?;
+                    }
+                }
+
+                first_pager
+                    .continuation_token()
+                    .expect("expected continuation token to be created after first page")
+            };
+            let mut rehydrated_pager = client
+                .list_secret_properties(None)?
+                .into_pages()
+                .with_continuation_token(rehydration_token);
+
+            while let Some(secret_page) = rehydrated_pager.try_next().await? {
+                let secrets = secret_page.into_model()?;
+                for secret in secrets.value {
+                    let _ = secret.resource_id()?;
+                }
+            }
+
+            Ok(())
+        },
+        ExpectedInstrumentation {
+            package_name: recording.var("CARGO_PKG_NAME", None),
+            // Don't use `recording.var` here in case the recording was made with a different package version.
+            package_version: env!("CARGO_PKG_VERSION").into(),
+            package_namespace: Some("KeyVault"),
+            api_calls: vec![
+                ExpectedApiInformation {
+                    api_name: Some("KeyVault.getSecrets"),
+                    api_children: vec![ExpectedRestApiSpan {
+                        api_verb: azure_core::http::Method::Get,
+                        ..Default::default()
+                    }],
+                    ..Default::default()
+                },
+                ExpectedApiInformation {
+                    api_name: Some("KeyVault.getSecrets"),
+                    api_children: vec![ExpectedRestApiSpan {
+                        api_verb: azure_core::http::Method::Get,
+                        is_wildcard: true,
+                        ..Default::default()
+                    }],
+                    ..Default::default()
+                },
+            ],
+        },
+    )
+    .await;
+
+    validate_result
 }