+++ sdk/core/typespec_client_core/CHANGELOG.md
@@ -1,9 +1,26 @@
 # Release History
 
-## 0.9.0 (Unreleased)
+## 0.10.0 (Unreleased)
 
 ### Features Added
 
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 0.9.1 (2025-11-09)
+
+### Other Changes
+
+- Increment version for re-release following a fix to publishing.
+
+## 0.9.0 (2025-11-07)
+
+### Features Added
+
+- Added `Context::to_owned()` to create a static lifetime copy of an existing `Context`.
 - Added `PipelineOptions::retry_status_codes` for configuring which status codes should trigger a retry.
 - Added `Response<T, F>::body(&self) -> &ResponseBody`.
 - Added `Response<T, F>::to_raw_response()` function to create a `RawResponse` from cloned data.
@@ -22,10 +39,6 @@
 - Renamed `BufResponseBody` to `AsyncResponseBody` so that `AsyncResponseBody` is to `ResponseBody` as `AsyncIterator` is to `Iterator`.
 - Renamed `Response<T, F>::into_body(self) -> Result<Response<T>>` to `into_model(self) -> Result<Response<T>>`. `into_body(self)` now returns a `ResponseBody`.
 - Renamed `xml::read_xml()` to `xml::from_xml()` congruent with `json::from_json()`.
-
-### Bugs Fixed
-
-### Other Changes
 
 ## 0.8.1 (2025-10-06)