+++ sdk/storage/azure_storage_blob/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Release History
 
-## 0.7.0 (Unreleased)
+## 0.8.0 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 0.7.0 (2025-11-11)
 
 ### Features Added